+++ src/tables/mod.rs
@@ -28,6 +28,7 @@ pub mod mvar;
 pub mod name;
 pub mod os2;
 pub mod post;
+pub mod sbix;
 pub mod stat;
 pub mod vhea;
 pub mod vmtx;